my_odoo_module/requirements.txt
# List any additional Python packages needed
odoo-addon-helper==1.0.0